[API-91] Extract buildApp + BuildAppOptions into api/build-app.ts.
